+++ console/src/dataSources/services/mssql/index.tsx
@@ -236,6 +236,7 @@ export const supportedFeatures: DeepRequired<SupportedFeaturesType> = {
     pool_timeout: false,
     connection_lifetime: false,
     ssl_certificates: false,
+    namingConvention: false,
   },
 };
 
@@ -292,7 +293,7 @@ WHERE
     'guest', 'INFORMATION_SCHEMA', 'sys',
     'db_owner', 'db_securityadmin', 'db_accessadmin',
     'db_backupoperator', 'db_ddladmin', 'db_datawriter',
-    'db_datareader', 'db_denydatawriter', 'db_denydatareader'
+    'db_datareader', 'db_denydatawriter', 'db_denydatareader', 'hdb_catalog'
   )
 ORDER BY
   s.name
@@ -932,7 +933,7 @@ WHERE
       'guest', 'INFORMATION_SCHEMA', 'sys',
       'db_owner', 'db_securityadmin', 'db_accessadmin',
       'db_backupoperator', 'db_ddladmin', 'db_datawriter',
-      'db_datareader', 'db_denydatawriter', 'db_denydatareader'
+      'db_datareader', 'db_denydatawriter', 'db_denydatareader', 'hdb_catalog'
     )
 ORDER BY
   table_schema,
@@ -977,7 +978,7 @@ WHERE
       'guest', 'INFORMATION_SCHEMA', 'sys',
       'db_owner', 'db_securityadmin', 'db_accessadmin',
       'db_backupoperator', 'db_ddladmin', 'db_datawriter',
-      'db_datareader', 'db_denydatawriter', 'db_denydatareader'
+      'db_datareader', 'db_denydatawriter', 'db_denydatareader', 'hdb_catalog'
     )
   ORDER BY
     s.name