+++ server/src-lib/Hasura/GraphQL/Schema/Action.hs
@@ -26,10 +26,10 @@ import Hasura.GraphQL.Parser
   )
 import Hasura.GraphQL.Parser qualified as P
 import Hasura.GraphQL.Parser.Class
-import Hasura.GraphQL.Parser.Constants qualified as G
 import Hasura.GraphQL.Parser.Internal.Parser qualified as P
 import Hasura.GraphQL.Schema.Backend
 import Hasura.GraphQL.Schema.Common
+import Hasura.Name qualified as Name
 import Hasura.Prelude
 import Hasura.RQL.IR.Action qualified as IR
 import Hasura.RQL.IR.Root qualified as IR
@@ -40,9 +40,6 @@ import Hasura.RQL.Types.Column
 import Hasura.RQL.Types.Common
 import Hasura.RQL.Types.CustomTypes
 import Hasura.RQL.Types.Relationships.Remote
-import Hasura.RQL.Types.SchemaCache
-import Hasura.RQL.Types.Source
-import Hasura.RQL.Types.SourceCustomization
 import Hasura.RQL.Types.Table
 import Hasura.SQL.AnyBackend qualified as AB
 import Hasura.SQL.Backend
@@ -54,14 +51,13 @@ import Language.GraphQL.Draft.Syntax qualified as G
 --   the field name and input arguments and a selectionset. The
 --   input argument and selectionset types are defined by the user.
 --
---
 -- > action_name(action_input_arguments) {
 -- >   col1: col1_type
 -- >   col2: col2_type
 -- > }
 actionExecute ::
   forall r m n.
-  MonadBuildSchema ('Postgres 'Vanilla) r m n =>
+  MonadBuildSchemaBase r m n =>
   AnnotatedCustomTypes ->
   ActionInfo ->
   m (Maybe (FieldParser n (IR.AnnActionExecution (IR.RemoteRelationshipField IR.UnpreparedValue))))
@@ -157,19 +153,19 @@ actionAsyncQuery objectTypes actionInfo = runMaybeT do
         let idField = P.selection_ idFieldName (Just idFieldDescription) actionIdParser $> IR.AsyncId
             createdAtField =
               P.selection_
-                G._created_at
+                Name._created_at
                 (Just "the time at which this action was created")
                 createdAtFieldParser
                 $> IR.AsyncCreatedAt
             errorsField =
               P.selection_
-                G._errors
+                Name._errors
                 (Just "errors related to the invocation")
                 errorsFieldParser
                 $> IR.AsyncErrors
             outputField =
               P.subselection_
-                G._output
+                Name._output
                 (Just "the output fields of this action")
                 actionOutputParser
                 <&> IR.AsyncOutput
@@ -187,6 +183,7 @@ actionAsyncQuery objectTypes actionInfo = runMaybeT do
       pure $ P.selection fieldName description actionIdInputField selectionSet <&> (,[])
 
   stringifyNum <- retrieve soStringifyNum
+  definitionsList <- lift $ mkDefinitionList outputObject
   pure $
     parserOutput
       <&> \(idArg, fields) ->
@@ -195,19 +192,57 @@ actionAsyncQuery objectTypes actionInfo = runMaybeT do
             _aaaqActionId = idArg,
             _aaaqOutputType = _adOutputType definition,
             _aaaqFields = fields,
-            _aaaqDefinitionList = mkDefinitionList outputObject,
+            _aaaqDefinitionList = definitionsList,
             _aaaqStringifyNum = stringifyNum,
             _aaaqForwardClientHeaders = forwardClientHeaders,
-            _aaaqSource = IR.getActionSourceInfo outputObject
+            _aaaqSource = getActionSourceInfo outputObject
           }
   where
     ActionInfo actionName (outputType, outputObject) definition permissions forwardClientHeaders comment = actionInfo
-    idFieldName = G._id
+    idFieldName = Name._id
     idFieldDescription = "the unique id of an action"
 
+    getActionSourceInfo :: AnnotatedOutputType -> IR.ActionSourceInfo ('Postgres 'Vanilla)
+    getActionSourceInfo = \case
+      AOTObject aot -> fromMaybe IR.ASINoSource $ listToMaybe do
+        AnnotatedTypeRelationship {..} <- _aotRelationships aot
+        pure $ IR.ASISource _atrSource _atrSourceConfig
+      AOTScalar _ -> IR.ASINoSource
+
+    mkDefinitionList :: AnnotatedOutputType -> m [(PGCol, ScalarType ('Postgres 'Vanilla))]
+    mkDefinitionList = \case
+      AOTScalar _ -> pure []
+      AOTObject AnnotatedObjectType {..} -> do
+        let fieldReferences = Map.unions $ map _atrFieldMapping _aotRelationships
+        for (toList _aotFields) \ObjectFieldDefinition {..} ->
+          (unsafePGCol . G.unName . unObjectFieldName $ _ofdName,)
+            <$> case Map.lookup _ofdName fieldReferences of
+              Nothing -> fieldTypeToScalarType $ snd _ofdType
+              Just columnInfo -> pure $ unsafePGColumnToBackend $ ciType columnInfo
+
+    -- warning: we don't support other backends than Postgres for async queries;
+    -- here, we fail if we encounter a non-Postgres scalar type
+    fieldTypeToScalarType :: AnnotatedObjectFieldType -> m PGScalarType
+    fieldTypeToScalarType = \case
+      AOFTEnum _ -> pure PGText
+      AOFTObject _ -> pure PGJSON
+      AOFTScalar annotatedScalar -> case annotatedScalar of
+        ASTReusedScalar _ scalar ->
+          case AB.unpackAnyBackend @('Postgres 'Vanilla) scalar of
+            Just pgScalar -> pure $ unwrapScalar pgScalar
+            Nothing -> throw500 "encountered non-Postgres scalar in async query actions"
+        ASTCustom ScalarTypeDefinition {..} ->
+          pure $
+            if
+                | _stdName == idScalar -> PGText
+                | _stdName == intScalar -> PGInteger
+                | _stdName == floatScalar -> PGFloat
+                | _stdName == stringScalar -> PGText
+                | _stdName == boolScalar -> PGBoolean
+                | otherwise -> PGJSON
+
 -- | Async action's unique id
-actionIdParser ::
-  MonadParse n => Parser 'Both n ActionId
+actionIdParser :: MonadParse n => Parser 'Both n ActionId
 actionIdParser = ActionId <$> P.uuid
 
 actionOutputFields ::
@@ -218,14 +253,13 @@ actionOutputFields ::
   HashMap G.Name AnnotatedObjectType ->
   m (Parser 'Output n (AnnotatedActionFields))
 actionOutputFields outputType annotatedObject objectTypes = do
-  let outputObject = _aotDefinition annotatedObject
-  scalarOrEnumOrObjectFields <- forM (toList $ _otdFields outputObject) outputFieldParser
-  relationshipFields <- traverse relationshipFieldParser $ _otdRelationships outputObject
-  outputTypeName <- P.mkTypename $ unObjectTypeName $ _otdName outputObject
+  scalarOrEnumOrObjectFields <- forM (toList $ _aotFields annotatedObject) outputFieldParser
+  relationshipFields <- traverse relationshipFieldParser $ _aotRelationships annotatedObject
+  outputTypeName <- P.mkTypename $ unObjectTypeName $ _aotName annotatedObject
   let allFieldParsers =
         scalarOrEnumOrObjectFields
           <> concat (catMaybes relationshipFields)
-      outputTypeDescription = _otdDescription outputObject
+      outputTypeDescription = _aotDescription annotatedObject
   pure $
     outputParserModifier outputType $
       P.selectionSet outputTypeName outputTypeDescription allFieldParsers
@@ -241,7 +275,7 @@ actionOutputFields outputType annotatedObject objectTypes = do
     outputFieldParser ::
       ObjectFieldDefinition (G.GType, AnnotatedObjectFieldType) ->
       m (FieldParser n (AnnotatedActionField))
-    outputFieldParser (ObjectFieldDefinition name _ description (gType, objectFieldType)) = memoizeOn 'actionOutputFields (_otdName $ _aotDefinition annotatedObject, name) do
+    outputFieldParser (ObjectFieldDefinition name _ description (gType, objectFieldType)) = memoizeOn 'actionOutputFields (_aotName annotatedObject, name) do
       case objectFieldType of
         AOFTScalar def ->
           wrapScalar $ customScalarParser def
@@ -259,12 +293,10 @@ actionOutputFields outputType annotatedObject objectTypes = do
               $> IR.ACFScalar fieldName
 
     relationshipFieldParser ::
-      TypeRelationship (TableInfo ('Postgres 'Vanilla)) (ColumnInfo ('Postgres 'Vanilla)) ->
+      AnnotatedTypeRelationship ->
       m (Maybe [FieldParser n (AnnotatedActionField)])
-    relationshipFieldParser (TypeRelationship relationshipName relType sourceName tableInfo fieldMapping) = runMaybeT do
-      sourceCache <- lift $ retrieve scSourceCache
-      sourceInfo <- hoistMaybe $ unsafeSourceInfo @('Postgres 'Vanilla) =<< Map.lookup sourceName sourceCache
-      relName <- hoistMaybe $ RelName <$> mkNonEmptyText (toTxt relationshipName)
+    relationshipFieldParser (AnnotatedTypeRelationship {..}) = runMaybeT do
+      relName <- hoistMaybe $ RelName <$> mkNonEmptyText (toTxt _atrName)
 
       --  `lhsJoinFields` is a map of `x: y`
       --  where 'x' is the 'reference name' of a join field, i.e, how a join
@@ -272,9 +304,9 @@ actionOutputFields outputType annotatedObject objectTypes = do
       --  while 'y' is the join field.
       --  In case of custom types, they are pretty much the same.
       --  In case of databases, 'y' could be a computed field with session variables etc.
-      let lhsJoinFields = Map.fromList [(FieldName $ G.unName k, k) | ObjectFieldName k <- Map.keys fieldMapping]
+      let lhsJoinFields = Map.fromList [(FieldName $ G.unName k, k) | ObjectFieldName k <- Map.keys _atrFieldMapping]
           joinMapping = Map.fromList $ do
-            (k, v) <- Map.toList fieldMapping
+            (k, v) <- Map.toList _atrFieldMapping
             let scalarType = case ciType v of
                   ColumnScalar scalar -> scalar
                   -- We don't currently allow enum types as fields of custom types so they should not appear here.
@@ -289,30 +321,17 @@ actionOutputFields outputType annotatedObject objectTypes = do
                     AB.mkAnyBackend @('Postgres 'Vanilla) $
                       RemoteSourceFieldInfo
                         { _rsfiName = relName,
-                          _rsfiType = relType,
-                          _rsfiSource = sourceName,
-                          _rsfiSourceConfig = _siConfiguration sourceInfo,
-                          _rsfiSourceCustomization = getSourceTypeCustomization $ _siCustomization sourceInfo,
-                          _rsfiTable = tableInfoName tableInfo,
+                          _rsfiType = _atrType,
+                          _rsfiSource = _atrSource,
+                          _rsfiSourceConfig = _atrSourceConfig,
+                          _rsfiSourceCustomization = _atrSourceCustomization,
+                          _rsfiTable = tableInfoName _atrTableInfo,
                           _rsfiMapping = joinMapping
                         }
               }
       RemoteRelationshipParserBuilder remoteRelationshipField <- retrieve scRemoteRelationshipParserBuilder
       remoteRelationshipFieldParsers <- MaybeT $ remoteRelationshipField remoteFieldInfo
       pure $ remoteRelationshipFieldParsers <&> fmap (IR.ACFRemote . IR.ActionRemoteRelationshipSelect lhsJoinFields)
-
-mkDefinitionList :: AnnotatedOutputType -> [(PGCol, ScalarType ('Postgres 'Vanilla))]
-mkDefinitionList (AOTScalar _) = []
-mkDefinitionList (AOTObject AnnotatedObjectType {..}) =
-  flip map (toList _otdFields) $ \ObjectFieldDefinition {..} ->
-    (unsafePGCol . G.unName . unObjectFieldName $ _ofdName,) $
-      case Map.lookup _ofdName fieldReferences of
-        Nothing -> fieldTypeToScalarType $ snd _ofdType
-        Just columnInfo -> unsafePGColumnToBackend $ ciType columnInfo
-  where
-    ObjectTypeDefinition {..} = _aotDefinition
-    fieldReferences =
-      Map.unions $ map _trFieldMapping _otdRelationships
 
 actionInputArguments ::
   forall r m n.
@@ -399,17 +418,25 @@ customScalarParser = \case
         | _stdName == stringScalar -> J.toJSON <$> P.string
         | _stdName == boolScalar -> J.toJSON <$> P.boolean
         | otherwise -> P.jsonScalar _stdName _stdDescription
-  ASTReusedScalar name pgScalarType ->
-    -- TODO: use column parser here instead
+  ASTReusedScalar name backendScalarType ->
     let schemaType = P.TNamed P.NonNullable $ P.Definition name Nothing P.TIScalar
+        backendScalarValidator =
+          AB.dispatchAnyBackend @Backend backendScalarType \(scalarType :: ScalarWrapper b) jsonInput -> do
+            -- We attempt to parse the value from JSON to validate it, but still
+            -- output it as JSON. On one hand this allows us to detect issues
+            -- ahead of time: if the value is not formatted correctly, we don't
+            -- send the action at all; on the other, it means we are at risk of
+            -- rejecting valid queries if our parser is more strict than the one
+            -- of the remote server. We do not parse scalars for remote servers
+            -- for that reason; we might want to reconsider this validation as
+            -- well.
+            void $
+              parseScalarValue @b (unwrapScalar scalarType) jsonInput
+                `onLeft` \e -> parseErrorWith ParseFailed $ qeError e
+            pure jsonInput
      in P.Parser
           { pType = schemaType,
-            pParser =
-              P.valueToJSON (P.toGraphQLType schemaType)
-                >=> either
-                  (parseErrorWith ParseFailed . qeError)
-                  (pure . scalarValueToJSON @('Postgres 'Vanilla))
-                  . parseScalarValue @('Postgres 'Vanilla) pgScalarType
+            pParser = P.valueToJSON (P.toGraphQLType schemaType) >=> backendScalarValidator
           }
 
 customEnumParser ::